+++ R/type-vctr.R
@@ -377,7 +377,7 @@ unique.vctrs_vctr <- function(x, incomparables = FALSE, ...) {
 
 #' @export
 duplicated.vctrs_vctr <- function(x, incomparables = FALSE, ...) {
-  vec_duplicate_id(x) != seq_along(x)
+  vec_duplicate_flg(x)
 }
 
 #' @export