+++ src/dictionary.c
@@ -141,6 +141,110 @@ SEXP vctrs_unique_loc(SEXP x) {
   return out;
 }
 
+SEXP vec_duplicate_loc_impl(SEXP x, R_len_t n);
+SEXP vec_duplicate_loc_first_impl(SEXP x, R_len_t n);
+
+SEXP vec_duplicate_loc(SEXP x, bool first) {
+  R_len_t n = vec_size(x);
+
+  x = PROTECT(vec_proxy_equal(x));
+  x = PROTECT(obj_maybe_translate_encoding(x, n));
+
+  SEXP out;
+
+  if (first) {
+    out = vec_duplicate_loc_first_impl(x, n);
+  } else {
+    out = vec_duplicate_loc_impl(x, n);
+  }
+
+  UNPROTECT(2);
+  return out;
+}
+
+SEXP vctrs_duplicate_loc(SEXP x, SEXP first) {
+  return vec_duplicate_loc(x, Rf_asLogical(first));
+}
+
+SEXP vec_duplicate_loc_impl(SEXP x, R_len_t n) {
+  int nprot = 0;
+
+  dictionary d;
+  dict_init(&d, x);
+  PROTECT_DICT(&d, &nprot);
+
+  struct growable g = new_growable(INTSXP, 256);
+  PROTECT_GROWABLE(&g, &nprot);
+
+  for (int i = 0; i < n; ++i) {
+    uint32_t hash = dict_hash_scalar(&d, i);
+
+    if (d.key[hash] == DICT_EMPTY) {
+      dict_put(&d, hash, i);
+    } else {
+      growable_push_int(&g, i + 1);
+    }
+  }
+
+  SEXP out = growable_values(&g);
+
+  UNPROTECT(nprot);
+  return out;
+}
+
+SEXP vec_duplicate_loc_first_impl(SEXP x, R_len_t n) {
+  int nprot = 0;
+
+  dictionary d;
+  dict_init(&d, x);
+  PROTECT_DICT(&d, &nprot);
+
+  SEXP val = PROTECT_N(Rf_allocVector(INTSXP, d.size), &nprot);
+  int* p_val = INTEGER(val);
+
+  // Store hash values so we don't have to look them up twice
+  int32_t* hashes = (int32_t*) R_alloc(n, sizeof(int32_t));
+
+  R_len_t n_out = 0;
+
+  for (int i = 0; i < n; ++i) {
+    int32_t hash = dict_hash_scalar(&d, i);
+    hashes[i] = hash;
+
+    if (d.key[hash] == DICT_EMPTY) {
+      dict_put(&d, hash, i);
+      p_val[hash] = 1;
+      continue;
+    }
+
+    ++p_val[hash];
+
+    if (p_val[hash] == 2) {
+      n_out += 2;
+    } else {
+      n_out += 1;
+    }
+  }
+
+  // Create output
+  SEXP out = PROTECT_N(Rf_allocVector(INTSXP, n_out), &nprot);
+  int* p_out = INTEGER(out);
+
+  int j = 0;
+
+  for (int i = 0; i < n; ++i) {
+    int32_t hash = hashes[i];
+
+    if (p_val[hash] != 1) {
+      p_out[j] = i + 1;
+      ++j;
+    }
+  }
+
+  UNPROTECT(nprot);
+  return out;
+}
+
 SEXP vctrs_duplicated_any(SEXP x) {
   bool out = duplicated_any(x);
   return Rf_ScalarLogical(out);
@@ -199,7 +303,7 @@ SEXP vctrs_n_distinct(SEXP x) {
   return Rf_ScalarInteger(d.used);
 }
 
-SEXP vctrs_id(SEXP x) {
+SEXP vctrs_first_loc(SEXP x) {
   int nprot = 0;
 
   R_len_t n = vec_size(x);
@@ -382,13 +486,61 @@ SEXP vctrs_count(SEXP x) {
   return out;
 }
 
-SEXP vctrs_duplicated(SEXP x) {
+SEXP vec_duplicate_flg_impl(SEXP x, R_len_t n);
+SEXP vec_duplicate_flg_first_impl(SEXP x, R_len_t n);
+
+SEXP vec_duplicate_flg(SEXP x, bool first) {
   int nprot = 0;
 
   R_len_t n = vec_size(x);
 
-  x = PROTECT_N(vec_proxy_equal(x), &nprot);
-  x = PROTECT_N(obj_maybe_translate_encoding(x, n), &nprot);
+  x = PROTECT(vec_proxy_equal(x));
+  x = PROTECT(obj_maybe_translate_encoding(x, n));
+
+  SEXP out;
+
+  if (first) {
+    out = vec_duplicate_flg_first_impl(x, n);
+  } else {
+    out = vec_duplicate_flg_impl(x, n);
+  }
+
+  UNPROTECT(2);
+  return out;
+}
+
+// [[ export() ]]
+SEXP vctrs_duplicate_flg(SEXP x, SEXP first) {
+  return vec_duplicate_flg(x, Rf_asLogical(first));
+}
+
+SEXP vec_duplicate_flg_impl(SEXP x, R_len_t n) {
+  int nprot = 0;
+
+  dictionary d;
+  dict_init(&d, x);
+  PROTECT_DICT(&d, &nprot);
+
+  SEXP out = PROTECT_N(Rf_allocVector(LGLSXP, n), &nprot);
+  int* p_out = LOGICAL(out);
+  memset(p_out, 0, n * sizeof(int));
+
+  for (int i = 0; i < n; ++i) {
+    int32_t hash = dict_hash_scalar(&d, i);
+
+    if (d.key[hash] == DICT_EMPTY) {
+      dict_put(&d, hash, i);
+    } else {
+      p_out[i] = 1;
+    }
+  }
+
+  UNPROTECT(nprot);
+  return out;
+}
+
+SEXP vec_duplicate_flg_first_impl(SEXP x, R_len_t n) {
+  int nprot = 0;
 
   dictionary d;
   dict_init(&d, x);
@@ -396,15 +548,20 @@ SEXP vctrs_duplicated(SEXP x) {
 
   SEXP val = PROTECT_N(Rf_allocVector(INTSXP, d.size), &nprot);
   int* p_val = INTEGER(val);
+  memset(p_val, 0, d.size * sizeof(int));
+
+  // Store hash values so we don't have to look them up twice
+  int32_t* hashes = (int32_t*) R_alloc(n, sizeof(int32_t));
 
   for (int i = 0; i < n; ++i) {
     int32_t hash = dict_hash_scalar(&d, i);
+    hashes[i] = hash;
 
     if (d.key[hash] == DICT_EMPTY) {
       dict_put(&d, hash, i);
-      p_val[hash] = 0;
+    } else {
+      p_val[hash] = 1;
     }
-    p_val[hash]++;
   }
 
   // Create output
@@ -412,8 +569,7 @@ SEXP vctrs_duplicated(SEXP x) {
   int* p_out = LOGICAL(out);
 
   for (int i = 0; i < n; ++i) {
-    int32_t hash = dict_hash_scalar(&d, i);
-    p_out[i] = p_val[hash] != 1;
+    p_out[i] = p_val[hashes[i]];
   }
 
   UNPROTECT(nprot);